data/requirements.txt
absl-py
aiohttp
argcomplete
argon2-cffi
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs
backcall
bleach
blinker==1.4
Bottleneck==1.3.2
brotlipy==0.7.0
cachetools
certifi==2021.5.30
cffi
chardet
charset-normalizer
click
coverage
cryptography
cycler==0.10.0
Cython
debugpy
decorator
defusedxml
dill
entrypoints==0.3
Flask
flatbuffers==1.12
fonttools==4.25.0
future==0.18.2
gast==0.3.3
gevent==21.8.0
google-auth
google-auth-oauthlib
google-pasta
googleapis-common-protos
greenlet==1.1.2
grpcio==1.32.0
h5py
idna
importlib-metadata
importlib-resources==5.2.2
ipykernel
ipython
ipython-genutils
ipywidgets
itsdangerous
jedi
Jinja2
joblib
jsonschema
jupyter==1.0.0
jupyter-client
jupyter-console
jupyter-core
jupyterlab-pygments
jupyterlab-widgets
Keras
Keras-Applications==1.0.8
Keras-Preprocessing
kiwisolver
Markdown
MarkupSafe
matplotlib
matplotlib-inline
mistune
mkl-fft==1.3.0
mkl-random
mkl-service==2.4.0
multidict
munkres==1.1.4
nbclient
nbconvert
nbformat
nest-asyncio
notebook
numexpr
numpy==1.19.5
oauthlib
olefile==0.46
opt-einsum
packaging
pandas==1.3.3
pandocfilters
parso
pexpect
pickleshare
Pillow
prometheus-client
promise
prompt-toolkit
protobuf==3.17.2
psutil
ptyprocess
pyasn1
pyasn1-modules==0.2.8
pycparser
Pygments
PyJWT
pyOpenSSL
pyparsing
pyrsistent
PySocks
python-dateutil
pytz
PyYAML==5.1.2
pyzmq
qtconsole
QtPy
requests
requests-oauthlib==1.3.0
rsa
scikit-learn
scipy
Send2Trash
six==1.15.0
tensorboard
tensorboard-plugin-wit==1.6.0
tensorflow
tensorflow-datasets
tensorflow-estimator==2.4.0
tensorflow-gpu==2.1.0rc2
tensorflow-metadata==0.14.0
termcolor==1.1.0
terminado==0.9.4
testpath
tfds-nightly==4.4.0.dev202109290107
threadpoolctl
tornado
tqdm
traitlets
typing-extensions==3.7.4.3
urllib3
uWSGI==2.0.18
wcwidth
webencodings==0.5.1
Werkzeug
widgetsnbextension==3.5.1
wrapt==1.12.1
yarl
zipp
zope.event==4.5.0
zope.interface==5.4.0
